requirements.txt
certifi==2018.11.29
Django==2.1.4
django-bootstrap4==0.0.7
djangorestframework==3.9.0
Pygments==2.3.1
pytz==2018.7
